requirements.txt
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.13
django-crispy-forms==1.6.0
django-filter==0.13.0
django-formtools==2.1
django-import-export==1.2.0
django-markdown-deux==1.0.5
django-otp==0.7.3
django-pagedown==0.1.1
django-phonenumber-field==1.3.0
django-rest-auth==0.9.5
django-two-factor-auth==1.9.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
markdown2==2.3.8
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
phonenumberslite==8.10.21
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.1.2
qrcode==6.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
twilio==6.32.0
urllib3==1.25.6
xlrd==1.2.0
xlwt==1.3.0

